+++ src/whl2conda/impl/prompt.py
@@ -15,6 +15,7 @@
 """
 Interactive prompt utilities.
 """
+
 from __future__ import annotations
 
 import io