+++ test/cli/test_config.py
@@ -15,6 +15,7 @@
 """
 Unit tests for `whl2conda config` CLI
 """
+
 from __future__ import annotations
 
 from pathlib import Path